+++ src/shared/lib/ExtendedRouterStore.spec.ts
@@ -49,7 +49,7 @@ describe('ExtendedRouterStore', () => {
         }, undefined, true);
 
 
-        assert.deepEqual(routingStore.query, {param3: 'cleared'},'removes param1');
+        assert.deepEqual(routingStore.location.query, {param3: 'cleared'},'removes param1');
         assert.deepEqual(routingStore.location.pathname,'/results');
 
     });
