+++ src/shared/lib/createQueryStore.ts
@@ -1,0 +1,30 @@
+import {CancerStudyQueryUrlParams, QueryStore} from "shared/components/query/QueryStore";
+import {ResultsViewTab} from "pages/resultsView/ResultsViewPageHelpers";
+import ResultsViewURLWrapper from "pages/resultsView/ResultsViewURLWrapper";
+
+export function createQueryStore(currentQuery?:any, clearUrl = true) {
+
+    const win:any = window;
+
+    const queryStore = new QueryStore(currentQuery);
+
+    queryStore.singlePageAppSubmitRoutine = function(query:CancerStudyQueryUrlParams) {
+
+        // normalize this
+        query.cancer_study_list = query.cancer_study_list || query.cancer_study_id;
+        delete query.cancer_study_id;
+
+        const tab = (queryStore.physicalStudyIdsInSelection.length > 1 && queryStore.geneIds.length === 1) ?
+            ResultsViewTab.CANCER_TYPES_SUMMARY : ResultsViewTab.ONCOPRINT;
+
+        const wrapper = new ResultsViewURLWrapper(win.routingStore);
+
+        wrapper.updateURL(query,`results/${tab}`, clearUrl, false);
+
+        wrapper.destroy();
+
+    };
+
+    return queryStore;
+
+}
